+++ detection/retinaface/rcnn/cython/gpu_nms.pyx
@@ -13,7 +13,7 @@ assert sizeof(int) == sizeof(np.int32_t)
 cdef extern from "gpu_nms.hpp":
     void _nms(np.int32_t*, int*, np.float32_t*, int, int, float, int)
 
-def gpu_nms(np.ndarray[np.float32_t, ndim=2] dets, np.float thresh,
+def gpu_nms(np.ndarray[np.float32_t, ndim=2] dets, np.float32 thresh,
             np.int32_t device_id=0):
     cdef int boxes_num = dets.shape[0]
     cdef int boxes_dim = dets.shape[1]
